few fixes

+++ src/main/java/cyanide3d/listener/CyanoListener.java
@@ -94,11 +94,11 @@ public class CyanoListener extends ListenerAdapter {
         if (user.isBot() || reactedUser.contains(user) || pins.isEmpty()) {
             return;
         }
-        if (event.retrieveMessage().complete().getAuthor().isBot() && event.getReaction().retrieveUsers().complete().stream().filter(user1 -> user1.isBot()).findAny() != null) {
+        if (event.retrieveMessage().complete().getAuthor().isBot() && event.getReaction().retrieveUsers().complete().stream().filter(user1 -> user1.isBot()).findAny() != null && pinService.getParseMessage().getId().equalsIgnoreCase(event.getMessageId())){
             String message = pins.get(pins.size()-1);
             user.openPrivateChannel().queue(privateChannel -> privateChannel.sendMessage(message).queue());
             ChannelManagmentService.getInstance().loggingChannel(event.getGuild()).sendMessage(user.getAsTag() + " **взял пин.**").queue();
-            pinService.setReactedUser(user);
+            //pinService.setReactedUser(user);
             pinService.removePin(pins.size()-1);
         }
     }
